migrate to BTs from long to BigDecimal

+++ converter-commons/src/test/java/it/infocert/eigor/converter/commons/cen2ubl/DocumentTotalsConverterTest.java
@@ -11,6 +11,8 @@ import org.jdom2.Element;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.*;
 
@@ -25,8 +27,8 @@ public class DocumentTotalsConverterTest {
     public void setUp() throws Exception {
         invoice = new BG0000Invoice();
         BG0022DocumentTotals totals = new BG0022DocumentTotals();
-        totals.getBT0106SumOfInvoiceLineNetAmount().add(new BT0106SumOfInvoiceLineNetAmount(100d));
-        totals.getBT0109InvoiceTotalAmountWithoutVat().add(new BT0109InvoiceTotalAmountWithoutVat(100d));
+        totals.getBT0106SumOfInvoiceLineNetAmount().add(new BT0106SumOfInvoiceLineNetAmount(BigDecimal.valueOf(100)));
+        totals.getBT0109InvoiceTotalAmountWithoutVat().add(new BT0109InvoiceTotalAmountWithoutVat(BigDecimal.valueOf(100)));
         invoice.getBG0022DocumentTotals().add(totals);
         invoice.getBT0005InvoiceCurrencyCode().add(new BT0005InvoiceCurrencyCode(Iso4217CurrenciesFundsCodes.EUR));
     }
@@ -75,13 +77,13 @@ public class DocumentTotalsConverterTest {
     private void enrichInvoiceWithBG0021() {
         BG0021DocumentLevelCharges bg0021 = new BG0021DocumentLevelCharges();
 
-        BT0099DocumentLevelChargeAmount bt0099 = new BT0099DocumentLevelChargeAmount(12.34);
+        BT0099DocumentLevelChargeAmount bt0099 = new BT0099DocumentLevelChargeAmount(BigDecimal.valueOf(12.34));
         bg0021.getBT0099DocumentLevelChargeAmount().add(bt0099);
 
-        BT0100DocumentLevelChargeBaseAmount bt0100 = new BT0100DocumentLevelChargeBaseAmount(34.56);
+        BT0100DocumentLevelChargeBaseAmount bt0100 = new BT0100DocumentLevelChargeBaseAmount(BigDecimal.valueOf(34.56));
         bg0021.getBT0100DocumentLevelChargeBaseAmount().add(bt0100);
 
-        BT0101DocumentLevelChargePercentage bt0101 = new BT0101DocumentLevelChargePercentage(56.78);
+        BT0101DocumentLevelChargePercentage bt0101 = new BT0101DocumentLevelChargePercentage(BigDecimal.valueOf(56.78));
         bg0021.getBT0101DocumentLevelChargePercentage().add(bt0101);
 
         BT0102DocumentLevelChargeVatCategoryCode bt0102 = new BT0102DocumentLevelChargeVatCategoryCode(Untdid5305DutyTaxFeeCategories.E);
